+++ lib/data/projects/work/projects-overview/workFrontEnd.ts
@@ -1,4 +1,5 @@
 import { iconsSkillsProjectsWorkSmartsuppWeb } from '@/lib/data/skills/skillsProjectsWork'
+
 import { PROJECT_ID } from '@/lib/utils/constants/ids/projectIds'
 import { Project } from '@/lib/utils/interfaces/interfaces'
