+++ lib/data/projects/work/projects-overview/workReact.ts
@@ -3,6 +3,7 @@ import {
   iconsSkillsProjectsWorkKooperativa,
   iconsSkillsProjectsWorkSmartsuppDashboard,
 } from '@/lib/data/skills/skillsProjectsWork'
+
 import { PROJECT_ID } from '@/lib/utils/constants/ids/projectIds'
 import { Project } from '@/lib/utils/interfaces/interfaces'
