+++ lib/data/websiteLinks.ts
@@ -3,7 +3,7 @@ import { TEXT } from '@/localization/english'
 
 import { WebsiteLinkItem } from '@/lib/utils/interfaces/interfaces'
 
-// TODO: refactor testId and id to utils/constants/ files
+// TODO: refactor testId and id to TESTING utils/constants/ files
 export const websiteLinks: WebsiteLinkItem[] = [
   { href: PAGES_URL.aboutMe, text: TEXT.aboutMe, testId: 'about-me-link', id: 'about-me' },
   { href: PAGES_URL.resume, text: TEXT.resume, testId: 'resume-link', id: 'resume' },
